split the codegen into a separate project so it runs first
