+++ components/kubernetes-api/pom.xml
@@ -154,31 +154,5 @@
 
       </build>
     </profile>
-
-    <profile>
-      <id>genkinds</id>
-
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.codehaus.mojo</groupId>
-            <artifactId>exec-maven-plugin</artifactId>
-            <version>${exec-maven-plugin.version}</version>
-            <executions>
-              <execution>
-                <goals>
-                  <goal>exec</goal>
-                </goals>
-              </execution>
-            </executions>
-            <configuration>
-              <mainClass>io.fabric8.kubernetes.api.GenerateKindToClassMap</mainClass>
-              <classpathScope>test</classpathScope>
-            </configuration>
-          </plugin>
-        </plugins>
-
-      </build>
-    </profile>
   </profiles>
 </project>
