+++ components/pom.xml
@@ -63,6 +63,7 @@
     <module>hubot-api</module>
     <module>jolokia-assertions</module>
     <module>karaf-features</module>
+    <module>kubernetes-codegen</module>
     <module>kubernetes-api</module>
     <module>kubernetes-assertions</module>
     <module>kubernetes-jolokia</module>
